// backend/src/pkg/dal/course/course.interface.ts
import { ICommon } from '../common/common.interface';

export interface ICourse extends ICommon {
  YEAR: string;
  SEMESTER: string;
  LEVEL_ID: string;
  YEAR_ID: string;
  ROUND_NUMBER: string;
  ROUND_NAME_TH: string;
  ROUND_NAME_EN: string;
  ROUND_START_DATE: string;
  ROUND_END_DATE: string;
  ROUND_ID: string;
  PROJECT_CODE: string;
  PROJECT_NAME_TH: string;
  PROJECT_NAME_EN: string;
  APPLICANT_CAN_UPLOAD_DOCUMENT: string;
  CURRICULUM_ID: string;
  PROJECT_ID: string;
  TCAS_PROGRAM_ID: string;
  STUDY_ADMIT_TYPE_ID: string;
  STUDY_PLAN_ID: string;
  STUDY_CLASSES_ID: string;
  SPECIFIC_NATION_ID: string;
  MASTER_REG_FACULTY_ID: string;
  MASTER_REG_CURRICULUM_ID: string;
  BRANCH_ID: string;
  URL: string;
  REMARK: string;
  PDF_EXISTS: string;
  CURRICULUM_FEE: string;
  START_APPLICATION_DATE: string;
  END_APPLICATION_DATE: string;
  START_APPLICATION_PAY_DATE: string;
  END_APPLICATION_PAY_DATE: string;
  START_CONFIRM_DATE: string;
  END_CONFIRM_DATE: string;
  START_CONFIRM_PAY_DATE: string;
  END_CONFIRM_PAY_DATE: string;
  NUMBER_OF_RANKS: string;
  CURRICULUM_FULL_NAME_EN: string;
  CURRICULUM_FULL_NAME_TH: string;
  CURRICULUM_SHORT_NAME_EN: string;
  CURRICULUM_SHORT_NAME_TH: string;
  CURRICULUM_YEAR: string;
  MASTER_REG_INTER: string;
  FACULTY_NAME_EN: string;
  FACULTY_NAME_TH: string;
  BRANCH_NAME_EN: string;
  BRANCH_NAME_TH: string;
  HOT_CONSTRAINT: string;
  STUDY_ADMIT_TYPE_NAME_TH: string;
  STUDY_ADMIT_TYPE_NAME_EN: string;
  STUDY_PLAN_NAME_TH: string;
  STUDY_PLAN_NAME_EN: string;
  STUDY_CLASSES_NAME_TH: string;
  STUDY_CLASSES_NAME_EN: string;
  SPECIFIC_NATION_NAME_TH: string;
  SPECIFIC_NATION_NAME_EN: string;
}